refactor other and policy
