+++ server/src/__tests__/normal/email-provider.tsx
@@ -1,5 +1,4 @@
 import { Database } from 'better-sqlite3'
-import { JSDOM } from 'jsdom'
 import {
   afterEach, beforeEach, describe, expect, Mock, test,
   vi,
@@ -10,12 +9,7 @@ import {
   routeConfig,
 } from 'configs'
 import app from 'index'
-import {
-  getApp,
-  getAuthorizeParams,
-  insertUsers,
-  postSignInRequest,
-} from 'tests/identity'
+import { insertUsers } from 'tests/identity'
 import {
   emailLogRecord,
   emailResponseMock,
@@ -23,7 +17,6 @@ import {
   migrate, mock,
   mockedKV,
 } from 'tests/mock'
-import { disableUser } from 'tests/util'
 
 let db: Database
 
@@ -36,151 +29,29 @@ afterEach(async () => {
   await mockedKV.empty()
 })
 
-describe(
-  'get /authorize-reset',
-  () => {
-    test(
-      'should show reset page',
-      async () => {
-        const appRecord = await getApp(db)
-        const params = await getAuthorizeParams(appRecord)
-
-        const res = await app.request(
-          `${routeConfig.IdentityRoute.AuthorizeReset}${params}`,
-          {},
-          mock(db),
-        )
-
-        const html = await res.text()
-        const dom = new JSDOM(html)
-        const document = dom.window.document
-        expect(document.getElementsByName('email').length).toBe(1)
-        expect(document.getElementsByName('password').length).toBe(1)
-        expect(document.getElementsByName('confirmPassword').length).toBe(1)
-        expect(document.getElementsByTagName('form').length).toBe(1)
-        expect(document.getElementsByTagName('select').length).toBe(1)
-      },
-    )
-
-    test(
-      'could disable locale selector',
-      async () => {
-        global.process.env.ENABLE_LOCALE_SELECTOR = false as unknown as string
-        const appRecord = await getApp(db)
-        const params = await getAuthorizeParams(appRecord)
-
-        const res = await app.request(
-          `${routeConfig.IdentityRoute.AuthorizeReset}${params}`,
-          {},
-          mock(db),
-        )
-
-        const html = await res.text()
-        const dom = new JSDOM(html)
-        const document = dom.window.document
-        expect(document.getElementsByTagName('select').length).toBe(0)
-        global.process.env.ENABLE_LOCALE_SELECTOR = true as unknown as string
-      },
-    )
-
-    test(
-      'should be blocked if not enable in config',
-      async () => {
-        global.process.env.ENABLE_PASSWORD_RESET = false as unknown as string
-        const appRecord = await getApp(db)
-        const params = await getAuthorizeParams(appRecord)
-
-        const res = await app.request(
-          `${routeConfig.IdentityRoute.AuthorizeReset}${params}`,
-          {},
-          mock(db),
-        )
-        expect(res.status).toBe(400)
-        global.process.env.ENABLE_PASSWORD_RESET = true as unknown as string
-      },
-    )
-  },
-)
-
-const testSendResetCode = async (route: string) => {
+const sendCorrectResetPasswordCodeReq = async () => {
   const body = {
     email: 'test@email.com',
     password: 'Password1!',
   }
 
   const res = await app.request(
-    route,
+    routeConfig.IdentityRoute.ResetPasswordCode,
     {
       method: 'POST', body: JSON.stringify(body),
     },
     mock(db),
   )
-  return res
+
+  const code = await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`) ?? ''
+
+  return {
+    res, code,
+  }
 }
 
 describe(
-  'post /reset-code',
-  () => {
-    test(
-      'should send reset code',
-      async () => {
-        const mockFetch = vi.fn(async () => {
-          return Promise.resolve({ ok: true })
-        })
-        global.fetch = mockFetch as Mock
-
-        await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
-        const json = await res.json()
-        expect(json).toStrictEqual({ success: true })
-
-        const code = await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`) ?? ''
-        expect(code.length).toBe(6)
-
-        expect(mockFetch).toBeCalledTimes(1)
-
-        const callArgs = mockFetch.mock.calls[0] as any[]
-        const body = (callArgs[1] as unknown as { body: string }).body
-        expect(callArgs[0]).toBe('https://api.sendgrid.com/v3/mail/send')
-        expect(body).toContain(code)
-
-        global.fetch = fetchMock
-      },
-    )
-
-    test(
-      'should return true if user is inactive',
-      async () => {
-        await insertUsers(db)
-        disableUser(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
-        const json = await res.json()
-        expect(json).toStrictEqual({ success: true })
-        expect(await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`)).toBeFalsy()
-      },
-    )
-
-    test(
-      'should throw error if no email provided',
-      async () => {
-        await insertUsers(db)
-
-        const res = await app.request(
-          routeConfig.IdentityRoute.ResetCode,
-          {
-            method: 'POST',
-            body: JSON.stringify({ password: 'Password1!' }),
-          },
-          mock(db),
-        )
-        expect(res.status).toBe(400)
-      },
-    )
-  },
-)
-
-describe(
-  'post /resend-reset-code',
+  'post /reset-password-code',
   () => {
     test(
       'should send reset code by Sendgrid',
@@ -196,13 +67,12 @@ describe(
         global.fetch = mockFetch as Mock
 
         await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const {
+          res, code,
+        } = await sendCorrectResetPasswordCodeReq()
         const json = await res.json()
         expect(json).toStrictEqual({ success: true })
-
-        const code = await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`) ?? ''
         expect(code.length).toBe(6)
-
         expect(mockFetch).toBeCalledTimes(1)
 
         const callArgs = mockFetch.mock.calls[0] as any[]
@@ -230,11 +100,11 @@ describe(
         global.fetch = mockFetch as Mock
 
         await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const {
+          res, code,
+        } = await sendCorrectResetPasswordCodeReq()
         const json = await res.json()
         expect(json).toStrictEqual({ success: true })
-
-        const code = await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`) ?? ''
         expect(code.length).toBe(6)
 
         expect(mockFetch).toBeCalledTimes(1)
@@ -259,11 +129,12 @@ describe(
         global.fetch = mockFetch as Mock
 
         await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const {
+          res, code,
+        } = await sendCorrectResetPasswordCodeReq()
         const json = await res.json()
         expect(json).toStrictEqual({ success: true })
 
-        const code = await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`) ?? ''
         expect(code.length).toBeFalsy()
         global.fetch = fetchMock
       },
@@ -285,11 +156,11 @@ describe(
         global.fetch = mockFetch as Mock
 
         await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const {
+          res, code,
+        } = await sendCorrectResetPasswordCodeReq()
         const json = await res.json()
         expect(json).toStrictEqual({ success: true })
-
-        const code = await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`) ?? ''
         expect(code.length).toBe(6)
 
         expect(mockFetch).toBeCalledTimes(1)
@@ -329,7 +200,7 @@ describe(
         global.fetch = mockFetch as Mock
 
         await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const { res } = await sendCorrectResetPasswordCodeReq()
         const json = await res.json()
         expect(json).toStrictEqual({ success: true })
 
@@ -362,11 +233,11 @@ describe(
         global.fetch = mockFetch as Mock
 
         await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const {
+          res, code,
+        } = await sendCorrectResetPasswordCodeReq()
         const json = await res.json()
         expect(json).toStrictEqual({ success: true })
-
-        const code = await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`) ?? ''
         expect(code.length).toBe(6)
 
         expect(mockFetch).toBeCalledTimes(1)
@@ -402,7 +273,7 @@ describe(
         global.fetch = mockFetch as Mock
 
         await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const { res } = await sendCorrectResetPasswordCodeReq()
         const json = await res.json()
         expect(json).toStrictEqual({ success: true })
 
@@ -437,11 +308,11 @@ describe(
         global.fetch = mockFetch as Mock
 
         await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const {
+          res, code,
+        } = await sendCorrectResetPasswordCodeReq()
         const json = await res.json()
         expect(json).toStrictEqual({ success: true })
-
-        const code = await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`) ?? ''
         expect(code.length).toBe(6)
 
         expect(mockFetch).toBeCalledTimes(1)
@@ -483,7 +354,7 @@ describe(
         global.fetch = mockFetch as Mock
 
         await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const { res } = await sendCorrectResetPasswordCodeReq()
         const json = await res.json()
         expect(json).toStrictEqual({ success: true })
 
@@ -516,7 +387,7 @@ describe(
 
         await insertUsers(db)
         const res = await app.request(
-          routeConfig.IdentityRoute.ResetCode,
+          routeConfig.IdentityRoute.ResetPasswordCode,
           {
             method: 'POST',
             body: JSON.stringify({
@@ -554,21 +425,21 @@ describe(
 
         await insertUsers(db)
 
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const { res } = await sendCorrectResetPasswordCodeReq()
         const json = await res.json()
         expect(json).toStrictEqual({ success: true })
         expect(await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetAttempts}-test@email.com`)).toBe('1')
 
-        const res1 = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const { res: res1 } = await sendCorrectResetPasswordCodeReq()
         const json1 = await res1.json()
         expect(json1).toStrictEqual({ success: true })
         expect(await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetAttempts}-test@email.com`)).toBe('2')
 
-        const res2 = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const { res: res2 } = await sendCorrectResetPasswordCodeReq()
         expect(res2.status).toBe(400)
 
         global.process.env.PASSWORD_RESET_EMAIL_THRESHOLD = 0 as unknown as string
-        const res3 = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const { res: res3 } = await sendCorrectResetPasswordCodeReq()
         expect(res3.status).toBe(200)
 
         global.process.env.PASSWORD_RESET_EMAIL_THRESHOLD = 5 as unknown as string
@@ -580,231 +451,10 @@ describe(
       async () => {
         global.process.env.SENDGRID_API_KEY = ''
         await insertUsers(db)
-        const res = await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
+        const { res } = await sendCorrectResetPasswordCodeReq()
         expect(res.status).toBe(400)
         expect(await res.text()).toBe(localeConfig.Error.NoEmailSender)
         global.process.env.SENDGRID_API_KEY = 'abc'
-      },
-    )
-  },
-)
-
-describe(
-  'post /authorize-reset',
-  () => {
-    test(
-      'should reset password',
-      async () => {
-        global.process.env.ACCOUNT_LOCKOUT_THRESHOLD = 1 as unknown as string
-        const appRecord = await getApp(db)
-        await insertUsers(db)
-
-        await postSignInRequest(
-          db,
-          appRecord,
-          { password: 'Password2!' },
-        )
-        expect(await mockedKV.get(`${adapterConfig.BaseKVKey.FailedLoginAttempts}-test@email.com`)).toBe('1')
-
-        await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
-
-        const body = {
-          email: 'test@email.com',
-          password: 'Password2!',
-          code: await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`),
-        }
-
-        const res = await app.request(
-          routeConfig.IdentityRoute.AuthorizeReset,
-          {
-            method: 'POST', body: JSON.stringify(body),
-          },
-          mock(db),
-        )
-        const json = await res.json()
-        expect(json).toStrictEqual({ success: true })
-
-        expect(await mockedKV.get(`${adapterConfig.BaseKVKey.FailedLoginAttempts}-test@email.com`)).toBeFalsy()
-
-        const signInRes = await postSignInRequest(
-          db,
-          appRecord,
-          { password: 'Password2!' },
-        )
-        expect(await signInRes.json()).toBeTruthy()
-        global.process.env.ACCOUNT_LOCKOUT_THRESHOLD = 5 as unknown as string
-      },
-    )
-
-    test(
-      'should throw error with wrong code',
-      async () => {
-        await insertUsers(db)
-
-        await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
-
-        const body = {
-          email: 'test@email.com',
-          password: 'Password2!',
-          code: 'abcdef',
-        }
-
-        const res = await app.request(
-          routeConfig.IdentityRoute.AuthorizeReset,
-          {
-            method: 'POST', body: JSON.stringify(body),
-          },
-          mock(db),
-        )
-        expect(res.status).toBe(400)
-        expect(await res.text()).toBe(localeConfig.Error.WrongCode)
-      },
-    )
-
-    test(
-      'should throw error when reset with same password',
-      async () => {
-        await insertUsers(db)
-
-        await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
-
-        const body = {
-          email: 'test@email.com',
-          password: 'Password1!',
-          code: await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`),
-        }
-
-        const res = await app.request(
-          routeConfig.IdentityRoute.AuthorizeReset,
-          {
-            method: 'POST', body: JSON.stringify(body),
-          },
-          mock(db),
-        )
-        expect(res.status).toBe(400)
-        expect(await res.text()).toBe(localeConfig.Error.RequireDifferentPassword)
-      },
-    )
-
-    test(
-      'should not reset if user is inactive',
-      async () => {
-        await insertUsers(db)
-
-        await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
-
-        disableUser(db)
-        const body = {
-          email: 'test@email.com',
-          password: 'Password2!',
-          code: await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`),
-        }
-
-        const res = await app.request(
-          routeConfig.IdentityRoute.AuthorizeReset,
-          {
-            method: 'POST', body: JSON.stringify(body),
-          },
-          mock(db),
-        )
-        expect(res.status).toBe(400)
-        expect(await res.text()).toBe(localeConfig.Error.UserDisabled)
-      },
-    )
-
-    test(
-      'should not reset if it is a wrong user',
-      async () => {
-        await insertUsers(db)
-
-        await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
-
-        disableUser(db)
-        const body = {
-          email: 'test1@email.com',
-          password: 'Password2!',
-          code: await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`),
-        }
-
-        const res = await app.request(
-          routeConfig.IdentityRoute.AuthorizeReset,
-          {
-            method: 'POST', body: JSON.stringify(body),
-          },
-          mock(db),
-        )
-        expect(res.status).toBe(404)
-        expect(await res.text()).toBe(localeConfig.Error.NoUser)
-      },
-    )
-
-    test(
-      'could disable account unlock by reset password',
-      async () => {
-        global.process.env.ACCOUNT_LOCKOUT_THRESHOLD = 1 as unknown as string
-        global.process.env.UNLOCK_ACCOUNT_VIA_PASSWORD_RESET = false as unknown as string
-
-        const appRecord = await getApp(db)
-        await insertUsers(db)
-
-        await postSignInRequest(
-          db,
-          appRecord,
-          { password: 'Password2!' },
-        )
-        expect(await mockedKV.get(`${adapterConfig.BaseKVKey.FailedLoginAttempts}-test@email.com`)).toBe('1')
-
-        await testSendResetCode(routeConfig.IdentityRoute.ResetCode)
-
-        const body = {
-          email: 'test@email.com',
-          password: 'Password2!',
-          code: await mockedKV.get(`${adapterConfig.BaseKVKey.PasswordResetCode}-1`),
-        }
-        await app.request(
-          routeConfig.IdentityRoute.AuthorizeReset,
-          {
-            method: 'POST', body: JSON.stringify(body),
-          },
-          mock(db),
-        )
-
-        expect(await mockedKV.get(`${adapterConfig.BaseKVKey.FailedLoginAttempts}-test@email.com`)).toBe('1')
-
-        global.process.env.ACCOUNT_LOCKOUT_THRESHOLD = 5 as unknown as string
-        global.process.env.UNLOCK_ACCOUNT_VIA_PASSWORD_RESET = true as unknown as string
-      },
-    )
-  },
-)
-
-describe(
-  'get /auth-code-expired',
-  () => {
-    test(
-      'should show auth code expired en view',
-      async () => {
-        const res = await app.request(
-          `${routeConfig.IdentityRoute.AuthCodeExpired}?locale=en`,
-          {},
-          mock(db),
-        )
-
-        const html = await res.text()
-        expect(html).toContain(localeConfig.authCodeExpired.msg.en)
-      },
-    )
-    test(
-      'should show auth code expired fr view',
-      async () => {
-        const res = await app.request(
-          `${routeConfig.IdentityRoute.AuthCodeExpired}?locale=fr`,
-          {},
-          mock(db),
-        )
-
-        const html = await res.text()
-        expect(html).toContain(localeConfig.authCodeExpired.msg.fr)
       },
     )
   },
