+++ server/src/configs/route.ts
@@ -20,14 +20,12 @@ export enum OauthRoute {
 export enum IdentityRoute {
   AuthorizeView = `${InternalRoute.Identity}/view/authorize`,
   ProcessView = `${InternalRoute.Identity}/view/process`,
-  AuthCodeExpiredView = `${InternalRoute.Identity}/view/auth-code-expired`,
   AuthorizeConsentInfo = `${InternalRoute.Identity}/consent-info`,
   AuthorizeMfaEnrollInfo = `${InternalRoute.Identity}/mfa-enroll-info`,
   AuthorizeOtpSetupInfo = `${InternalRoute.Identity}/otp-setup-info`,
   AuthorizeOtpMfaInfo = `${InternalRoute.Identity}/otp-mfa-info`,
   AuthorizeSmsMfaInfo = `${InternalRoute.Identity}/sms-mfa-info`,
   AuthorizePasskeyEnrollInfo = `${InternalRoute.Identity}/passkey-enroll-info`,
-  ManagePasskeyInfo = `${InternalRoute.Identity}/manage-passkey-info`,
   AuthCodeExpired = `${InternalRoute.Identity}/auth-code-expired`,
   AuthorizePassword = `${InternalRoute.Identity}/authorize-password`,
   AuthorizeAccount = `${InternalRoute.Identity}/authorize-account`,
@@ -46,16 +44,18 @@ export enum IdentityRoute {
   AuthorizeGoogle = `${InternalRoute.Identity}/authorize-google`,
   AuthorizeFacebook = `${InternalRoute.Identity}/authorize-facebook`,
   AuthorizeGitHub = `${InternalRoute.Identity}/authorize-github`,
-  AuthorizeReset = `${InternalRoute.Identity}/authorize-reset`,
-  ResetCode = `${InternalRoute.Identity}/reset-code`,
   Logout = `${InternalRoute.Identity}/logout`,
+  // Policy
   ChangePassword = `${InternalRoute.Identity}/change-password`,
-  ChangeEmail = `${InternalRoute.Identity}/change-email`,
   ChangeEmailCode = `${InternalRoute.Identity}/change-email-code`,
+  ChangeEmail = `${InternalRoute.Identity}/change-email`,
   ResetMfa = `${InternalRoute.Identity}/reset-mfa`,
   ManagePasskey = `${InternalRoute.Identity}/manage-passkey`,
   UpdateInfo = `${InternalRoute.Identity}/update-info`,
   // Other
   VerifyEmailView = `${InternalRoute.Identity}/view/verify-email`,
   VerifyEmail = `${InternalRoute.Identity}/verify-email`,
+  ResetPasswordCode = `${InternalRoute.Identity}/reset-password-code`,
+  ResetPassword = `${InternalRoute.Identity}/reset-password`,
+  AuthCodeExpiredView = `${InternalRoute.Identity}/view/auth-code-expired`,
 }
