+++ server/src/routes/identity.tsx
@@ -22,11 +22,6 @@ identityRoutes.get(
 )
 
 identityRoutes.get(
-  routeConfig.IdentityRoute.AuthCodeExpiredView,
-  identityHandler.getAuthCodeExpiredView,
-)
-
-identityRoutes.get(
   routeConfig.IdentityRoute.AuthorizeConsentInfo,
   identityHandler.getAuthorizeConsentInfo,
 )
@@ -218,26 +213,6 @@ identityRoutes.get(
   identityHandler.getAuthorizeGithub,
 )
 
-identityRoutes.get(
-  routeConfig.IdentityRoute.AuthorizeReset,
-  configMiddleware.enablePasswordReset,
-  identityHandler.getAuthorizeReset,
-)
-
-identityRoutes.post(
-  routeConfig.IdentityRoute.AuthorizeReset,
-  setupMiddleware.validOrigin,
-  configMiddleware.enablePasswordReset,
-  identityHandler.postAuthorizeReset,
-)
-
-identityRoutes.post(
-  routeConfig.IdentityRoute.ResetCode,
-  setupMiddleware.validOrigin,
-  configMiddleware.enablePasswordReset,
-  identityHandler.postResetCode,
-)
-
 identityRoutes.post(
   routeConfig.IdentityRoute.Logout,
   authMiddleware.spa,
@@ -249,12 +224,14 @@ identityRoutes.get(
   identityHandler.getAuthCodeExpired,
 )
 
-identityRoutes.get(
-  routeConfig.IdentityRoute.ChangePassword,
-  configMiddleware.enableChangePasswordPolicy,
-  identityHandler.getChangePassword,
-)
-
+/**
+ * Policy
+ * - ChangePassword
+ * - ChangeEmail
+ * - ResetMfa
+ * - ManagePasskey
+ * - UpdateInfo
+ */
 identityRoutes.post(
   routeConfig.IdentityRoute.ChangePassword,
   setupMiddleware.validOrigin,
@@ -262,16 +239,10 @@ identityRoutes.post(
   identityHandler.postChangePassword,
 )
 
-identityRoutes.get(
-  routeConfig.IdentityRoute.ChangeEmail,
-  configMiddleware.enableChangeEmailPolicy,
-  identityHandler.getChangeEmail,
-)
-
 identityRoutes.post(
   routeConfig.IdentityRoute.ChangeEmailCode,
   configMiddleware.enableChangeEmailPolicy,
-  identityHandler.postVerificationCode,
+  identityHandler.postChangeEmailCode,
 )
 
 identityRoutes.post(
@@ -280,22 +251,10 @@ identityRoutes.post(
   identityHandler.postChangeEmail,
 )
 
-identityRoutes.get(
-  routeConfig.IdentityRoute.ResetMfa,
-  configMiddleware.enableResetMfaPolicy,
-  identityHandler.getResetMfa,
-)
-
 identityRoutes.post(
   routeConfig.IdentityRoute.ResetMfa,
   configMiddleware.enableResetMfaPolicy,
   identityHandler.postResetMfa,
-)
-
-identityRoutes.get(
-  routeConfig.IdentityRoute.ManagePasskeyInfo,
-  configMiddleware.enableManagePasskeyPolicy,
-  identityHandler.getManagePasskeyInfo,
 )
 
 identityRoutes.get(
@@ -316,12 +275,6 @@ identityRoutes.delete(
   identityHandler.deleteManagePasskey,
 )
 
-identityRoutes.get(
-  routeConfig.IdentityRoute.UpdateInfo,
-  configMiddleware.enableUpdateInfoPolicy,
-  identityHandler.getUpdateInfo,
-)
-
 identityRoutes.post(
   routeConfig.IdentityRoute.UpdateInfo,
   configMiddleware.enableUpdateInfoPolicy,
@@ -332,6 +285,7 @@ identityRoutes.post(
  * Other
  * - VerifyEmail
  * - ResetPassword
+ * - AuthCodeExpired
  */
 
 identityRoutes.get(
@@ -345,4 +299,23 @@ identityRoutes.post(
   setupMiddleware.validOrigin,
   configMiddleware.enableEmailVerification,
   identityHandler.postVerifyEmail,
+)
+
+identityRoutes.post(
+  routeConfig.IdentityRoute.ResetPasswordCode,
+  setupMiddleware.validOrigin,
+  configMiddleware.enablePasswordReset,
+  identityHandler.postResetPasswordCode,
+)
+
+identityRoutes.post(
+  routeConfig.IdentityRoute.ResetPassword,
+  setupMiddleware.validOrigin,
+  configMiddleware.enablePasswordReset,
+  identityHandler.postResetPassword,
+)
+
+identityRoutes.get(
+  routeConfig.IdentityRoute.AuthCodeExpiredView,
+  identityHandler.getAuthCodeExpiredView,
 )
